src/main/java/org/duyphung/vocamemo/repositories/RoleRepository.java
package org.duyphung.vocamemo.repositories;

import org.duyphung.vocamemo.entities.RoleEntity;
import org.springframework.data.jpa.repository.JpaRepository;
import org.springframework.data.jpa.repository.Query;
import org.springframework.data.repository.query.Param;
import org.springframework.stereotype.Repository;

@Repository
public interface RoleRepository extends JpaRepository<RoleEntity, Long> {

    public RoleEntity findRoleByName(String role);

    @Query(value = "select * from role where role.id= (select role_id from users_roles where user_id = :id)", nativeQuery = true)
    public RoleEntity findRoleByUser(@Param("id") long id);
}